+++ Hints.py
@@ -1433,16 +1433,23 @@ def get_junk_hint(spoiler: Spoiler, world: World, checked: set[str]) -> HintRetu
 
 def get_important_check_hint(spoiler: Spoiler, world: World, checked: set[str]) -> HintReturn:
     top_level_locations = []
+    empty_dungeons = [dungeon for dungeon in world.precompleted_dungeons if world.precompleted_dungeons[dungeon]]
     for location in world.get_filled_locations():
-        if (HintArea.at(location).text(world.settings.clearer_hints) not in top_level_locations
-                and (HintArea.at(location).text(world.settings.clearer_hints) + ' Important Check') not in checked
-                and HintArea.at(location) != HintArea.ROOT):
-            top_level_locations.append(HintArea.at(location).text(world.settings.clearer_hints))
-    hint_loc = random.choice(top_level_locations)
+        hint_area = HintArea.at(location)
+        if (
+            hint_area not in top_level_locations
+            and hint_area not in checked
+            and hint_area != HintArea.ROOT
+            and hint_area.dungeon_name not in empty_dungeons # prevent pre-completed dungeons from being hinted
+            and not location.locked # prevent areas with unshuffled checks from being hinted
+        ):
+            top_level_locations.append(hint_area)
+    if not top_level_locations:
+        return None
+    hint_area = random.choice(top_level_locations)
     item_count = 0
     for location in world.get_filled_locations():
-        region = HintArea.at(location).text(world.settings.clearer_hints)
-        if region == hint_loc:
+        if HintArea.at(location) == hint_area:
             if (location.item.majoritem
                 # exclude locked items
                 and not location.locked
@@ -1464,7 +1471,7 @@ def get_important_check_hint(spoiler: Spoiler, world: World, checked: set[str]) 
                     or world.shuffle_ganon_bosskey == 'tokens' or world.shuffle_ganon_bosskey == 'hearts'))):
                 item_count = item_count + 1
 
-    checked.add(hint_loc + ' Important Check')
+    checked.add(hint_area)
 
     if item_count == 0:
         numcolor = 'Red'
@@ -1477,7 +1484,7 @@ def get_important_check_hint(spoiler: Spoiler, world: World, checked: set[str]) 
     else:
         numcolor = 'Green'
 
-    return GossipText('%s has #%d# major item%s.' % (hint_loc, item_count, "s" if item_count != 1 else ""), ['Green', numcolor]), None
+    return GossipText('%s has #%d# major item%s.' % (hint_area.text(world.settings.clearer_hints), item_count, "s" if item_count != 1 else ""), ['Green', numcolor]), None
 
 
 hint_func: dict[str, HintFunc | BarrenFunc] = {